Introduce AiO HotPlug v1.0
Signed-off-by: Shoaib0597 <Shoaib0595@gmail.com>

An All in One HotPlug for Octa-Core big.LITTLE SoCs. This HotPlug has been developed by me i.e., Shoaib Anwar a.k.a. Shoaib0597.

This HotPlug aims to provide all of the functionality which a user wants from a HotPlug, ranging from normal offline/online of Cores to completely offlining entire Clusters. As the name suggests, this HotPlug is a complete package.

Conflicts:
	drivers/soc/qcom/Kconfig
	drivers/soc/qcom/Makefile